% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ARI.R
\name{ARI}
\alias{ARI}
\title{Adjusted Rand index}
\usage{
ARI(x, y)
}
\arguments{
\item{x}{vector with clustering, matrix with hot-one-encoding of the
clustering, or a list of clusterings (in vector or matrix form)}

\item{y}{as x}
}
\value{
ARI (scalar of vector)
}
\description{
ARI to compare two clusterings or to compare two entire lists of clusterings
}
\examples{
x <- c(1,1,2,2,3,3)
y <- c(1,1,1,2,2,2)
ARI(x,y)

x <- matrix(0, 3, 6)
x[1,1] <- x[1,2] <- x[2,3] <- x[2,4] <- x[3,5] <- x[3,6] <- 1
y <- matrix(0, 2, 6)
y[1,1] <- y[1,2] <- y[1,3] <- y[2,4] <- y[2,5] <- y[2,6] <- 1
ARI(x,y)

X <- list(c(1,1,2,2,3,3), rep(1,10))
Y <- list(c(1,1,1,2,2,2), rep(1:2,each=5))
ARI(X,Y)
}
