% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matchBlocks.R
\name{graphonL2norm}
\alias{graphonL2norm}
\title{(squared) L2-norm of the graphons associated with two stochastic block model parameters}
\usage{
graphonL2norm(theta1, theta2)
}
\arguments{
\item{theta1}{a stochastic block model parameter}

\item{theta2}{a stochastic block model parameter}
}
\value{
(squared) L2-norm of the graphons associated with two stochastic block model parameters
}
\description{
(squared) L2-norm of the graphons associated with two stochastic block model parameters
}
\examples{
theta1 <- list(pi=c(.5,.5), gamma=matrix((1:4)/8,2,2))
theta2 <- list(pi=c(.5,.5), gamma=matrix(4:1/8,2,2))
graphonL2norm(theta1, theta2)
}
