% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mukherjee.R
\name{moments}
\alias{moments}
\title{Computation of graph moments of a network}
\usage{
moments(A, k = 3)
}
\arguments{
\item{A}{adjacency matrix}

\item{k}{order of the largest graph moments to be considered}
}
\value{
vector with the first k (normalized) graph moments of the network A
}
\description{
Computation of graph moments of a network
}
\examples{
param <- list(R = 500, alpha = .04, beta = .02, deltaIn = 100, deltaOut = 100)
A <- sampleDPA(param)
moments(A)
}
