% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/directedPrefAttachment.R
\name{sampleDPA}
\alias{sampleDPA}
\title{generation of a network of the directed preferential attachment (DPA) model}
\usage{
sampleDPA(param)
}
\arguments{
\item{param}{list with the following elements: $R (= number of iterations),
$alpha, $beta , $deltaIn, $deltaOut (parameters of the DPA model)}
}
\value{
adjacency matrix of generated network
}
\description{
generation of a network of the directed preferential attachment (DPA) model
}
\examples{
param <- list(R = 500, alpha = .04, beta = .02, deltaIn = 100, deltaOut = 100)
A <- sampleDPA(param)
A
}
